+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - [2026-05-17 11:06] home/godmode/config/prefill.json
 - [2026-05-17 11:08] home/godmode/lib/parsers.py
 - [2026-05-17 11:10] home/godmode/lib/encoder.py
+- [2026-05-17 11:13] home/godmode/engine/racer.py